connecting the database
it took so long to connect the postgresql database! I almost lost the hope and was thinking to make a different layout for managing the orders. The issues came when I had node js v19 but the strapi was supporting 18 and below. Then I learned nvm, it is a damn nice thing. Finally end of the day the database got connected, hopefully from tomorrow I'll be working on logics and adding content.